A few external server/docker improvements (#255)
* Fix lockup when external server not present on startup

* Clean up restart command

Disposing checks if the server actually needs stopping.
Using dispose prevents trying to stop the connection when the underlying
socket is already gone. VSCode sends a shutdown command and marks the
client as `Stopping`. But Aquiring the connection never resolves or errors
resulting in the language server never actually restarting

* Offer users to try connecting again for external clients

Co-authored-by: Fred Snyder <fsnyder@castwide.com>